Dish out a function to combinations of variables

Evaluates a two argument function on subsets of a data frame by evaluating each combination of columns or subsets.

dish(
    data,
    f,
    left = NULL,
    right = NULL,
    each_left = TRUE,
    each_right = TRUE,
    bind = FALSE,
    ...
)

Arguments

data
A data.frame.

f
Any function that takes a vector and/or data.frame in the first two arguments.

left
Variables to be used in the first argument of f. If NULL (default), all variables except those in right are used. Has tidyselect::vars_select capabilities.

right
Variables to be used in the second argument of f. If NULL (default), all variables except those in left are used. Has tidyselect::vars_select capabilities.

each_left
Should each left variable be separately evaluated in f? Defaults to TRUE. If FALSE, left is entered into f as a data.frame.

each_right
Should each right variable be separately evaluated in f? Defaults to TRUE. If FALSE, right is entered into f as a data.frame.

bind
Should results be binded into a single data.frame? Defaults to FALSE.

...
Additional arguments to be passed to f.

Value

A list (if bind = FALSE) or a tibble (if bind = TRUE) with the results of f evaluated on data subsets.

Examples

require(tidyverse)

#1) Default uses every variable on both sides
heart_disease %>%
    select_if(
        is.numeric
    ) %>%
    dish(
        f = cor
    )

#> $Age
#> $Age$Age
#> [1] 1
#> 
#> $Age$BP
#> [1] 0.2849459
#> 
#> $Age$Cholesterol
#> [1] 0.2089503
#> 
#> $Age$MaximumHR
#> [1] -0.3938058
#> 
#> 
#> $BP
#> $BP$Age
#> [1] 0.2849459
#> 
#> $BP$BP
#> [1] 1
#> 
#> $BP$Cholesterol
#> [1] 0.1301201
#> 
#> $BP$MaximumHR
#> [1] -0.04535088
#> 
#> 
#> $Cholesterol
#> $Cholesterol$Age
#> [1] 0.2089503
#> 
#> $Cholesterol$BP
#> [1] 0.1301201
#> 
#> $Cholesterol$Cholesterol
#> [1] 1
#> 
#> $Cholesterol$MaximumHR
#> [1] -0.003431832
#> 
#> 
#> $MaximumHR
#> $MaximumHR$Age
#> [1] -0.3938058
#> 
#> $MaximumHR$BP
#> [1] -0.04535088
#> 
#> $MaximumHR$Cholesterol
#> [1] -0.003431832
#> 
#> $MaximumHR$MaximumHR
#> [1] 1
#> 
#> 

#2) Simple regression of Age and BP on each variable
heart_disease %>%
    dish(
        f =
            function(y, x) {
                mod <- lm(y ~ x)
                tibble(
                    Parameter = names(mod$coef),
                    Estimate = mod$coef
                )
            },
        left = c("Age", "BP"),
        bind = TRUE
    )

#> # A tibble: 32 x 4
#>    .left .right      Parameter         Estimate
#>    <chr> <chr>       <chr>                <dbl>
#>  1 Age   Sex         (Intercept)        55.7   
#>  2 Age   Sex         xMale              -1.89  
#>  3 Age   ChestPain   (Intercept)        55.9   
#>  4 Age   ChestPain   xAtypical angina   -4.51  
#>  5 Age   ChestPain   xNon-anginal pain  -2.17  
#>  6 Age   ChestPain   xAsymptomatic      -0.147 
#>  7 Age   Cholesterol (Intercept)        45.4   
#>  8 Age   Cholesterol x                   0.0365
#>  9 Age   BloodSugar  (Intercept)        54.0   
#> 10 Age   BloodSugar  xTRUE               3.01  
#> # … with 22 more rows

#3) Multiple regression with Age, BP on all variables simultaneously
heart_disease %>%
    dish(
        f =
            function(y, x) {
                mod <- lm(y ~ ., data = x)
                tibble(
                    Parameter = names(mod$coef),
                    Estimate = mod$coef
                )
            },
        left = c("Age", "BP"),
        each_right = FALSE,
        bind = TRUE
    )

#> # A tibble: 20 x 3
#>    .left Parameter                 Estimate
#>    <chr> <chr>                        <dbl>
#>  1 Age   (Intercept)                72.9   
#>  2 Age   SexMale                    -2.16  
#>  3 Age   ChestPainAtypical angina   -3.74  
#>  4 Age   ChestPainNon-anginal pain  -2.80  
#>  5 Age   ChestPainAsymptomatic      -3.10  
#>  6 Age   Cholesterol                 0.0322
#>  7 Age   BloodSugarTRUE              2.80  
#>  8 Age   MaximumHR                  -0.153 
#>  9 Age   ExerciseInducedAnginaYes   -1.91  
#> 10 Age   HeartDiseaseYes             2.23  
#> 11 BP    (Intercept)               130.    
#> 12 BP    SexMale                    -4.61  
#> 13 BP    ChestPainAtypical angina  -11.8   
#> 14 BP    ChestPainNon-anginal pain -11.0   
#> 15 BP    ChestPainAsymptomatic     -11.8   
#> 16 BP    Cholesterol                 0.0328
#> 17 BP    BloodSugarTRUE              8.08  
#> 18 BP    MaximumHR                   0.0166
#> 19 BP    ExerciseInducedAnginaYes    0.665 
#> 20 BP    HeartDiseaseYes             6.88