Divide a data frame into a list

Separate a data.frame into a list of any depth by one or more stratification columns whose levels become the names.

divide(
    data,
    ...,
    depth = Inf,
    remove = TRUE,
    drop = TRUE,
    sep = "|"
)

Arguments

data
Any data.frame.

...
Selection of columns to split by. See ?dplyr::select for details.

depth
Depth to split the list to. Defaults to Inf. See details for more information.

remove
Should the stratfication columns be removed? Defaults to TRUE.

drop
Should unused combinations of stratification variables be dropped? Defaults to TRUE.

sep
String to separate values of each stratification variable by. Defaults to "|". Only used when the number of stratification columns exceeds the desired depth.

Details

For the depth, use positive integers to move from the root and negative integers to move from the leaves. The maximum depth will be used for integers larger than such.

Value

A list with tibble::tibble objects at the leaves.

Examples

require(tidyverse)

#1) Unquoted
heart_disease %>%
    divide(
        Sex
    )

#> $Female
#> # A tibble: 97 x 8
#>      Age ChestPain    BP Cholesterol BloodSugar MaximumHR ExerciseInduced…
#>    <dbl> <fct>     <dbl>       <dbl> <lgl>          <dbl> <fct>           
#>  1    41 Atypical…   130         204 FALSE            172 No              
#>  2    62 Asymptom…   140         268 FALSE            160 No              
#>  3    57 Asymptom…   120         354 FALSE            163 Yes             
#>  4    56 Atypical…   140         294 FALSE            153 No              
#>  5    48 Non-angi…   130         275 FALSE            139 No              
#>  6    58 Typical …   150         283 TRUE             162 No              
#>  7    50 Non-angi…   120         219 FALSE            158 No              
#>  8    58 Non-angi…   120         340 FALSE            172 No              
#>  9    66 Typical …   150         226 FALSE            114 No              
#> 10    69 Typical …   140         239 FALSE            151 No              
#> # … with 87 more rows, and 1 more variable: HeartDisease <fct>
#> 
#> $Male
#> # A tibble: 206 x 8
#>      Age ChestPain    BP Cholesterol BloodSugar MaximumHR ExerciseInduced…
#>    <dbl> <fct>     <dbl>       <dbl> <lgl>          <dbl> <fct>           
#>  1    63 Typical …   145         233 TRUE             150 No              
#>  2    67 Asymptom…   160         286 FALSE            108 Yes             
#>  3    67 Asymptom…   120         229 FALSE            129 Yes             
#>  4    37 Non-angi…   130         250 FALSE            187 No              
#>  5    56 Atypical…   120         236 FALSE            178 No              
#>  6    63 Asymptom…   130         254 FALSE            147 No              
#>  7    53 Asymptom…   140         203 TRUE             155 Yes             
#>  8    57 Asymptom…   140         192 FALSE            148 No              
#>  9    56 Non-angi…   130         256 TRUE             142 Yes             
#> 10    44 Atypical…   120         263 FALSE            173 No              
#> # … with 196 more rows, and 1 more variable: HeartDisease <fct>
#> 

#2) Using select helpers
heart_disease %>%
    divide(
        matches("^S")
    )

#> $Female
#> # A tibble: 97 x 8
#>      Age ChestPain    BP Cholesterol BloodSugar MaximumHR ExerciseInduced…
#>    <dbl> <fct>     <dbl>       <dbl> <lgl>          <dbl> <fct>           
#>  1    41 Atypical…   130         204 FALSE            172 No              
#>  2    62 Asymptom…   140         268 FALSE            160 No              
#>  3    57 Asymptom…   120         354 FALSE            163 Yes             
#>  4    56 Atypical…   140         294 FALSE            153 No              
#>  5    48 Non-angi…   130         275 FALSE            139 No              
#>  6    58 Typical …   150         283 TRUE             162 No              
#>  7    50 Non-angi…   120         219 FALSE            158 No              
#>  8    58 Non-angi…   120         340 FALSE            172 No              
#>  9    66 Typical …   150         226 FALSE            114 No              
#> 10    69 Typical …   140         239 FALSE            151 No              
#> # … with 87 more rows, and 1 more variable: HeartDisease <fct>
#> 
#> $Male
#> # A tibble: 206 x 8
#>      Age ChestPain    BP Cholesterol BloodSugar MaximumHR ExerciseInduced…
#>    <dbl> <fct>     <dbl>       <dbl> <lgl>          <dbl> <fct>           
#>  1    63 Typical …   145         233 TRUE             150 No              
#>  2    67 Asymptom…   160         286 FALSE            108 Yes             
#>  3    67 Asymptom…   120         229 FALSE            129 Yes             
#>  4    37 Non-angi…   130         250 FALSE            187 No              
#>  5    56 Atypical…   120         236 FALSE            178 No              
#>  6    63 Asymptom…   130         254 FALSE            147 No              
#>  7    53 Asymptom…   140         203 TRUE             155 Yes             
#>  8    57 Asymptom…   140         192 FALSE            148 No              
#>  9    56 Non-angi…   130         256 TRUE             142 Yes             
#> 10    44 Atypical…   120         263 FALSE            173 No              
#> # … with 196 more rows, and 1 more variable: HeartDisease <fct>
#> 

#3) Reduce the depth
heart_disease %>%
    divide(
        Sex,
        HeartDisease,
        depth = 1
    ) %>%
    length() #Each combinations

#> [1] 4

#4) Keep columns in result
heart_disease %>%
    divide(
        Sex,
        remove = FALSE
    )

#> $Female
#> # A tibble: 97 x 9
#>      Age Sex   ChestPain    BP Cholesterol BloodSugar MaximumHR ExerciseInduced…
#>    <dbl> <fct> <fct>     <dbl>       <dbl> <lgl>          <dbl> <fct>           
#>  1    41 Fema… Atypical…   130         204 FALSE            172 No              
#>  2    62 Fema… Asymptom…   140         268 FALSE            160 No              
#>  3    57 Fema… Asymptom…   120         354 FALSE            163 Yes             
#>  4    56 Fema… Atypical…   140         294 FALSE            153 No              
#>  5    48 Fema… Non-angi…   130         275 FALSE            139 No              
#>  6    58 Fema… Typical …   150         283 TRUE             162 No              
#>  7    50 Fema… Non-angi…   120         219 FALSE            158 No              
#>  8    58 Fema… Non-angi…   120         340 FALSE            172 No              
#>  9    66 Fema… Typical …   150         226 FALSE            114 No              
#> 10    69 Fema… Typical …   140         239 FALSE            151 No              
#> # … with 87 more rows, and 1 more variable: HeartDisease <fct>
#> 
#> $Male
#> # A tibble: 206 x 9
#>      Age Sex   ChestPain    BP Cholesterol BloodSugar MaximumHR ExerciseInduced…
#>    <dbl> <fct> <fct>     <dbl>       <dbl> <lgl>          <dbl> <fct>           
#>  1    63 Male  Typical …   145         233 TRUE             150 No              
#>  2    67 Male  Asymptom…   160         286 FALSE            108 Yes             
#>  3    67 Male  Asymptom…   120         229 FALSE            129 Yes             
#>  4    37 Male  Non-angi…   130         250 FALSE            187 No              
#>  5    56 Male  Atypical…   120         236 FALSE            178 No              
#>  6    63 Male  Asymptom…   130         254 FALSE            147 No              
#>  7    53 Male  Asymptom…   140         203 TRUE             155 Yes             
#>  8    57 Male  Asymptom…   140         192 FALSE            148 No              
#>  9    56 Male  Non-angi…   130         256 TRUE             142 Yes             
#> 10    44 Male  Atypical…   120         263 FALSE            173 No              
#> # … with 196 more rows, and 1 more variable: HeartDisease <fct>
#> 

#5) Move from leaves
heart_disease %>%
    divide(
        Sex,
        HeartDisease,
        ChestPain,
        depth = -1
    ) %>%
    str()

#> List of 2
#>  $ Female:List of 7
#>   ..$ No|Typical angina   :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	4 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:4] 58 66 69 60
#>   .. ..$ BP                   : num [1:4] 150 150 140 150
#>   .. ..$ Cholesterol          : num [1:4] 283 226 239 240
#>   .. ..$ BloodSugar           : logi [1:4] TRUE FALSE FALSE FALSE
#>   .. ..$ MaximumHR            : num [1:4] 162 114 151 171
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 1 1 1 1
#>   ..$ No|Atypical angina  :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	16 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:16] 41 56 71 41 45 55 54 45 63 50 ...
#>   .. ..$ BP                   : num [1:16] 130 140 160 105 130 135 132 112 140 120 ...
#>   .. ..$ Cholesterol          : num [1:16] 204 294 302 198 234 250 288 160 195 244 ...
#>   .. ..$ BloodSugar           : logi [1:16] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..$ MaximumHR            : num [1:16] 172 153 162 168 175 161 159 138 179 162 ...
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 1 1 1 1 1 1 2 1 1 1 ...
#>   ..$ Yes|Atypical angina :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	2 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:2] 58 57
#>   .. ..$ BP                   : num [1:2] 136 130
#>   .. ..$ Cholesterol          : num [1:2] 319 236
#>   .. ..$ BloodSugar           : logi [1:2] TRUE FALSE
#>   .. ..$ MaximumHR            : num [1:2] 152 174
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 1 1
#>   ..$ No|Non-anginal pain :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	34 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:34] 48 50 58 65 46 54 65 65 51 53 ...
#>   .. ..$ BP                   : num [1:34] 130 120 120 140 142 135 155 160 140 128 ...
#>   .. ..$ Cholesterol          : num [1:34] 275 219 340 417 177 304 269 360 308 216 ...
#>   .. ..$ BloodSugar           : logi [1:34] FALSE FALSE FALSE TRUE FALSE TRUE ...
#>   .. ..$ MaximumHR            : num [1:34] 139 158 172 157 160 170 148 151 142 115 ...
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 1 1 1 1 2 1 1 1 1 1 ...
#>   ..$ Yes|Non-anginal pain:Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1 obs. of  6 variables:
#>   .. ..$ Age                  : num 62
#>   .. ..$ BP                   : num 130
#>   .. ..$ Cholesterol          : num 263
#>   .. ..$ BloodSugar           : logi FALSE
#>   .. ..$ MaximumHR            : num 97
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 1
#>   ..$ No|Asymptomatic     :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	18 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:18] 57 53 53 57 35 62 42 58 62 45 ...
#>   .. ..$ BP                   : num [1:18] 120 130 138 128 138 124 102 100 140 138 ...
#>   .. ..$ Cholesterol          : num [1:18] 354 264 234 303 183 209 265 248 394 236 ...
#>   .. ..$ BloodSugar           : logi [1:18] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..$ MaximumHR            : num [1:18] 163 143 160 159 182 163 122 122 157 152 ...
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 2 1 1 1 1 1 1 1 1 2 ...
#>   ..$ Yes|Asymptomatic    :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	22 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:22] 62 65 61 51 62 60 61 43 63 56 ...
#>   .. ..$ BP                   : num [1:22] 140 150 130 130 160 150 145 132 150 200 ...
#>   .. ..$ Cholesterol          : num [1:22] 268 225 330 305 164 258 307 341 407 288 ...
#>   .. ..$ BloodSugar           : logi [1:22] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..$ MaximumHR            : num [1:22] 160 114 169 142 145 157 146 136 154 133 ...
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 1 1 1 2 1 1 2 2 1 2 ...
#>  $ Male  :List of 8
#>   ..$ No|Typical angina   :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	12 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:12] 63 64 40 51 34 52 52 42 59 69 ...
#>   .. ..$ BP                   : num [1:12] 145 110 140 125 118 118 152 148 178 160 ...
#>   .. ..$ Cholesterol          : num [1:12] 233 211 199 213 182 186 298 244 270 234 ...
#>   .. ..$ BloodSugar           : logi [1:12] TRUE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..$ MaximumHR            : num [1:12] 150 144 178 125 174 190 178 178 145 131 ...
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 1 2 2 2 1 1 1 1 1 1 ...
#>   ..$ Yes|Typical angina  :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	7 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:7] 65 59 59 38 61 59 45
#>   .. ..$ BP                   : num [1:7] 138 170 160 120 134 134 110
#>   .. ..$ Cholesterol          : num [1:7] 282 288 273 231 234 204 264
#>   .. ..$ BloodSugar           : logi [1:7] TRUE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..$ MaximumHR            : num [1:7] 174 159 125 182 145 162 132
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 1 1 1 2 1 1 1
#>   ..$ No|Atypical angina  :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	25 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:25] 56 44 49 44 48 52 52 54 41 44 ...
#>   .. ..$ BP                   : num [1:25] 120 120 130 130 130 120 134 108 135 120 ...
#>   .. ..$ Cholesterol          : num [1:25] 236 263 266 219 245 325 201 309 203 220 ...
#>   .. ..$ BloodSugar           : logi [1:25] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..$ MaximumHR            : num [1:25] 178 173 171 188 180 172 158 156 132 170 ...
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ Yes|Atypical angina :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	7 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:7] 48 58 62 66 54 57 57
#>   .. ..$ BP                   : num [1:7] 110 120 120 160 192 124 154
#>   .. ..$ Cholesterol          : num [1:7] 229 284 281 246 283 261 232
#>   .. ..$ BloodSugar           : logi [1:7] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..$ MaximumHR            : num [1:7] 168 160 103 120 195 141 164
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 1 1 1 2 1 1 1
#>   ..$ No|Non-anginal pain :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	34 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:34] 37 52 57 44 61 59 51 53 54 54 ...
#>   .. ..$ BP                   : num [1:34] 130 172 150 130 150 150 110 130 125 150 ...
#>   .. ..$ Cholesterol          : num [1:34] 250 199 168 233 243 212 175 197 273 232 ...
#>   .. ..$ BloodSugar           : logi [1:34] FALSE TRUE FALSE FALSE TRUE TRUE ...
#>   .. ..$ MaximumHR            : num [1:34] 187 162 174 179 137 157 123 152 152 165 ...
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 1 1 1 2 2 1 1 1 1 1 ...
#>   ..$ Yes|Non-anginal pain:Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	17 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:17] 56 58 64 58 50 60 46 68 49 57 ...
#>   .. ..$ BP                   : num [1:17] 130 132 140 112 140 140 150 180 120 128 ...
#>   .. ..$ Cholesterol          : num [1:17] 256 224 335 230 233 185 231 274 188 229 ...
#>   .. ..$ BloodSugar           : logi [1:17] TRUE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..$ MaximumHR            : num [1:17] 142 173 158 165 163 155 147 150 139 150 ...
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 2 1 1 1 1 1 1 2 1 1 ...
#>   ..$ No|Asymptomatic     :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	21 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:21] 57 54 43 59 42 65 45 66 48 45 ...
#>   .. ..$ BP                   : num [1:21] 140 140 150 135 140 120 104 120 122 115 ...
#>   .. ..$ Cholesterol          : num [1:21] 192 239 247 234 226 177 208 302 222 260 ...
#>   .. ..$ BloodSugar           : logi [1:21] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..$ MaximumHR            : num [1:21] 148 160 171 161 178 140 148 151 186 185 ...
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 1 1 1 1 1 1 2 1 1 1 ...
#>   ..$ Yes|Asymptomatic    :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	83 obs. of  6 variables:
#>   .. ..$ Age                  : num [1:83] 67 67 63 53 60 40 60 43 57 55 ...
#>   .. ..$ BP                   : num [1:83] 160 120 130 140 130 110 117 120 150 132 ...
#>   .. ..$ Cholesterol          : num [1:83] 286 229 254 203 206 167 230 177 276 353 ...
#>   .. ..$ BloodSugar           : logi [1:83] FALSE FALSE FALSE TRUE FALSE FALSE ...
#>   .. ..$ MaximumHR            : num [1:83] 108 129 147 155 132 114 160 120 112 132 ...
#>   .. ..$ ExerciseInducedAngina: Factor w/ 2 levels "No","Yes": 2 2 1 2 2 2 2 2 2 2 ...