Stratify a data frame, apply a function to each strata, and collect results

Calls "divide" to separate a data frame into one or more stratified frames, applies a function to each strata, and optionally collects the results and returns strata to their original columns.

stratiply(
    data,
    strata,
    f,
    delimiter = "|",
    bind = FALSE,
    separate = TRUE,
    ...
)

Arguments

data
Any data.frame.

strata
A vector of column names or positions, or a tidyselect::vars_select helper.

f
Any function that takes a data.frame as an argument.

delimiter
See the sep argument in "divide". Also used for salvaging original columns.

bind
Should the results be binded back to a single data.frame? Defaults to FALSE.

separate
If bind = TRUE, should the stratification variables be separated back to their original columns? Defaults to TRUE.

...
Additional arguments passed to f.

Value

If bind = FALSE, a named list with the results of f for each strata. Otherwise the results are binded to a single data.frame.

Examples

require(tidyverse)

#Return a list
heart_disease %>%
    stratiply(
        strata = c("Sex", "HeartDisease"),
        f = head
    )

#> $`Female|No`
#> # A tibble: 6 x 7
#>     Age ChestPain         BP Cholesterol BloodSugar MaximumHR ExerciseInducedAn…
#>   <dbl> <fct>          <dbl>       <dbl> <lgl>          <dbl> <fct>             
#> 1    41 Atypical angi…   130         204 FALSE            172 No                
#> 2    57 Asymptomatic     120         354 FALSE            163 Yes               
#> 3    56 Atypical angi…   140         294 FALSE            153 No                
#> 4    48 Non-anginal p…   130         275 FALSE            139 No                
#> 5    58 Typical angina   150         283 TRUE             162 No                
#> 6    50 Non-anginal p…   120         219 FALSE            158 No                
#> 
#> $`Male|No`
#> # A tibble: 6 x 7
#>     Age ChestPain         BP Cholesterol BloodSugar MaximumHR ExerciseInducedAn…
#>   <dbl> <fct>          <dbl>       <dbl> <lgl>          <dbl> <fct>             
#> 1    63 Typical angina   145         233 TRUE             150 No                
#> 2    37 Non-anginal p…   130         250 FALSE            187 No                
#> 3    56 Atypical angi…   120         236 FALSE            178 No                
#> 4    57 Asymptomatic     140         192 FALSE            148 No                
#> 5    44 Atypical angi…   120         263 FALSE            173 No                
#> 6    52 Non-anginal p…   172         199 TRUE             162 No                
#> 
#> $`Female|Yes`
#> # A tibble: 6 x 7
#>     Age ChestPain       BP Cholesterol BloodSugar MaximumHR ExerciseInducedAngi…
#>   <dbl> <fct>        <dbl>       <dbl> <lgl>          <dbl> <fct>               
#> 1    62 Asymptomatic   140         268 FALSE            160 No                  
#> 2    65 Asymptomatic   150         225 FALSE            114 No                  
#> 3    61 Asymptomatic   130         330 FALSE            169 No                  
#> 4    51 Asymptomatic   130         305 FALSE            142 Yes                 
#> 5    62 Asymptomatic   160         164 FALSE            145 No                  
#> 6    60 Asymptomatic   150         258 FALSE            157 No                  
#> 
#> $`Male|Yes`
#> # A tibble: 6 x 7
#>     Age ChestPain         BP Cholesterol BloodSugar MaximumHR ExerciseInducedAn…
#>   <dbl> <fct>          <dbl>       <dbl> <lgl>          <dbl> <fct>             
#> 1    67 Asymptomatic     160         286 FALSE            108 Yes               
#> 2    67 Asymptomatic     120         229 FALSE            129 Yes               
#> 3    63 Asymptomatic     130         254 FALSE            147 No                
#> 4    53 Asymptomatic     140         203 TRUE             155 Yes               
#> 5    56 Non-anginal p…   130         256 TRUE             142 Yes               
#> 6    48 Atypical angi…   110         229 FALSE            168 No                
#> 

#Separated strata column
heart_disease %>%
    stratiply(
        strata = c("Sex", "HeartDisease"),
        f = head,
        bind = TRUE
    )

#> # A tibble: 24 x 9
#>    Sex   HeartDisease   Age ChestPain    BP Cholesterol BloodSugar MaximumHR
#>    <chr> <chr>        <dbl> <fct>     <dbl>       <dbl> <lgl>          <dbl>
#>  1 Fema… No              41 Atypical…   130         204 FALSE            172
#>  2 Fema… No              57 Asymptom…   120         354 FALSE            163
#>  3 Fema… No              56 Atypical…   140         294 FALSE            153
#>  4 Fema… No              48 Non-angi…   130         275 FALSE            139
#>  5 Fema… No              58 Typical …   150         283 TRUE             162
#>  6 Fema… No              50 Non-angi…   120         219 FALSE            158
#>  7 Male  No              63 Typical …   145         233 TRUE             150
#>  8 Male  No              37 Non-angi…   130         250 FALSE            187
#>  9 Male  No              56 Atypical…   120         236 FALSE            178
#> 10 Male  No              57 Asymptom…   140         192 FALSE            148
#> # … with 14 more rows, and 1 more variable: ExerciseInducedAngina <fct>

#Custom function
heart_disease %>%
    stratiply(
        strata = matches("^(S|H)"),
        f = function(x)
            x %>%
            select_if(is.numeric) %>%
            map(mean, na.rm = TRUE),
        bind = TRUE,
        separate = TRUE
    )

#> # A tibble: 4 x 6
#>   Sex    HeartDisease   Age    BP Cholesterol MaximumHR
#>   <chr>  <chr>        <dbl> <dbl>       <dbl>     <dbl>
#> 1 Female No            54.6  129.        257.      154.
#> 2 Male   No            51.0  130.        232.      162.
#> 3 Female Yes           59.1  147.        276.      143.
#> 4 Male   Yes           56.1  132.        246.      138.