Compute association statistics between columns of a data frame

Evaluate a base::list of scalar functions on any number of "response" columns by any number of "predictor" columns

univariate_associations(
    data,
    f,
    responses,
    predictors
)

Arguments

data
A base::data.frame.

f
A function or a base::list of functions (preferably named) that take a vector as input in the first two arguments and return a scalar.

responses
A vector of quoted/unquoted columns, positions, and/or tidyselect::select_helpers to be evaluated as the first argument. See cheese::dish.

predictors
A vector of quoted/unquoted columns, positions, and/or tidyselect::select_helpers to be evaluated as the second argument. See cheese::dish.

Value

A tibble::tibble with the response/predictor columns down the rows and the results of the f across the columns. The names of the result columns will be the names provided in f.

Examples

#Make a list of functions to evaluate
f <-
  list(

    #Compute a univariate p-value
    `P-value` =
      function(y, x) {
        if(some_type(x, c("factor", "character"))) {

          p <- fisher.test(factor(y), factor(x), simulate.p.value = TRUE)$p.value

        } else {

          p <- kruskal.test(x, factor(y))$p.value

        }

        ifelse(p < 0.001, "<0.001", as.character(round(p, 2)))

      },

    #Compute difference in AIC model between null model and one predictor model
    `AIC Difference` =
      function(y, x) {

        glm(factor(y)~1, family = "binomial")$aic -
          glm(factor(y)~x, family = "binomial")$aic

      }
  )

#Choose a couple binary outcomes
heart_disease %>%
  univariate_associations(
    f = f,
    responses = c(ExerciseInducedAngina, HeartDisease)
  )

#> # A tibble: 14 x 4
#>    response              predictor   `P-value` `AIC Difference`
#>    <chr>                 <chr>       <chr>                <dbl>
#>  1 ExerciseInducedAngina Age         0.14                 0.567
#>  2 ExerciseInducedAngina Sex         0.01                 4.72 
#>  3 ExerciseInducedAngina ChestPain   <0.001              64.1  
#>  4 ExerciseInducedAngina BP          0.38                -0.740
#>  5 ExerciseInducedAngina Cholesterol 0.14                -0.875
#>  6 ExerciseInducedAngina BloodSugar  0.66                -1.80 
#>  7 ExerciseInducedAngina MaximumHR   <0.001              42.5  
#>  8 HeartDisease          Age         <0.001              13.4  
#>  9 HeartDisease          Sex         <0.001              22.0  
#> 10 HeartDisease          ChestPain   <0.001              80.1  
#> 11 HeartDisease          BP          0.03                 4.95 
#> 12 HeartDisease          Cholesterol 0.04                 0.206
#> 13 HeartDisease          BloodSugar  0.66                -1.81 
#> 14 HeartDisease          MaximumHR   <0.001              55.1  

#Use a subset of predictors
heart_disease %>%
  univariate_associations(
    f = f,
    responses = c(ExerciseInducedAngina, HeartDisease),
    predictors = c(Age, BP)
  )

#> # A tibble: 4 x 4
#>   response              predictor `P-value` `AIC Difference`
#>   <chr>                 <chr>     <chr>                <dbl>
#> 1 ExerciseInducedAngina Age       0.14                 0.567
#> 2 ExerciseInducedAngina BP        0.38                -0.740
#> 3 HeartDisease          Age       <0.001              13.4  
#> 4 HeartDisease          BP        0.03                 4.95 

#Numeric predictors only
heart_disease %>%
  univariate_associations(
    f = f,
    responses = c(ExerciseInducedAngina, HeartDisease),
    predictors = is.numeric
  )

#> # A tibble: 8 x 4
#>   response              predictor   `P-value` `AIC Difference`
#>   <chr>                 <chr>       <chr>                <dbl>
#> 1 ExerciseInducedAngina Age         0.14                 0.567
#> 2 ExerciseInducedAngina BP          0.38                -0.740
#> 3 ExerciseInducedAngina Cholesterol 0.14                -0.875
#> 4 ExerciseInducedAngina MaximumHR   <0.001              42.5  
#> 5 HeartDisease          Age         <0.001              13.4  
#> 6 HeartDisease          BP          0.03                 4.95 
#> 7 HeartDisease          Cholesterol 0.04                 0.206
#> 8 HeartDisease          MaximumHR   <0.001              55.1  

#Make a HTML table with spanned header
heart_disease %>%
  univariate_associations(
    f = f,
    responses = c(ExerciseInducedAngina, HeartDisease),
    predictors = is.numeric
  ) %>%

  #Spread responses across the columns
  stretch(
    key = response,
    value = names(f)
  ) %>%

  #Span the header in a table
  grable(
    at = -predictor
  )

#> <table class="table" style="margin-left: auto; margin-right: auto;">
#>  <thead>
#> <tr>
#> <th style="border-bottom:hidden" colspan="1"></th>
#> <th style="border-bottom:hidden; padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="2"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">ExerciseInducedAngina</div></th>
#> <th style="border-bottom:hidden; padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="2"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">HeartDisease</div></th>
#> </tr>
#>   <tr>
#>    <th style="text-align:left;"> predictor </th>
#>    <th style="text-align:left;"> P-value </th>
#>    <th style="text-align:right;"> AIC Difference </th>
#>    <th style="text-align:left;"> P-value </th>
#>    <th style="text-align:right;"> AIC Difference </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> Age </td>
#>    <td style="text-align:left;"> 0.14 </td>
#>    <td style="text-align:right;"> 0.5673180 </td>
#>    <td style="text-align:left;"> &lt;0.001 </td>
#>    <td style="text-align:right;"> 13.4465871 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> BP </td>
#>    <td style="text-align:left;"> 0.38 </td>
#>    <td style="text-align:right;"> -0.7398564 </td>
#>    <td style="text-align:left;"> 0.03 </td>
#>    <td style="text-align:right;"> 4.9541705 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Cholesterol </td>
#>    <td style="text-align:left;"> 0.14 </td>
#>    <td style="text-align:right;"> -0.8750976 </td>
#>    <td style="text-align:left;"> 0.04 </td>
#>    <td style="text-align:right;"> 0.2055648 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> MaximumHR </td>
#>    <td style="text-align:left;"> &lt;0.001 </td>
#>    <td style="text-align:right;"> 42.4780494 </td>
#>    <td style="text-align:left;"> &lt;0.001 </td>
#>    <td style="text-align:right;"> 55.0800334 </td>
#>   </tr>
#> </tbody>
#> </table>