Create a custom univariate summary for a dataset

Produces a formatted table of univariate summary statistics with options allowing for stratification by 1 or more variables, computing of custom summary/association statistics, custom string templates for results, etc.

univariate_table(
    data,
    strata = NULL,
    associations = NULL,
    numeric_summary = c(Summary = "median (q1, q3)"),
    categorical_summary = c(Summary = "count (percent%)"),
    other_summary = c(Summary = "unique"),
    all_summary = NULL,
    evaluate = FALSE,
    add_n = FALSE,
    order = NULL,
    labels = NULL,
    levels = NULL,
    format = c("html", "latex", "markdown", "pandoc", "none"),
    variableName = "Variable",
    levelName = "Level",
    na_string = "(missing)",
    strata_sep = "/",
    summary_strata_sep = "_",
    fill_blanks = "",
    caption = "",
    ...
)

Arguments

data
A data.frame to summarise.

strata
A formula specifying one or more stratification variables. LHS variables go to rows, RHS variables go to columns. Defaults to NULL.

associations
A named list of functions to evaluate with column strata and each variable. Defaults to NULL.

numeric_summary
A (preferably named) character vector containing string templates of how results for numeric data should be presented. See details for a list of what is available by default. Defaults to c(Summary = "median (q1, q3)").

categorical_summary
A (preferably named) character vector containing string templates of how results for categorical data should be presented. See details for a list of what is available by default. Defaults to c(Summary = "count (percent%)").

other_summary
A (preferably named) character vector containing string templates of how results for non-numeric and non-categorical data should be presented. See details for a list of what is available by default. Defaults to c(Summary = "unique").

all_summary
A (preferably named) character vector containing string templates of additional results for all variables should be presented. See details for a list of what is available by default. Defaults to NULL.

evaluate
Should the results of the string templates be evaluated as an R expression after filled with their values? See "absorb" for details. Defaults to FALSE.

add_n
Should the sample size for each stratfication level be added to the result? Defaults to FALSE.

order
Character vector of 1 or more variables to reorder the result by from top to bottom. If NULL (default), the result is sorted according to names(data).

labels
Named character vector for re-labeling variables in the result. Defaults to NULL.

levels
Named list of character vectors for re-labeling factor levels in the result. Defaults to NULL.

format
The format that the result should be rendered as. Must be one of c("html", "latex", "markdown", "pandoc", "none"). Defaults to "html".

variableName
Header for the variable column in the result. Defaults to "Variable".

levelName
Header for the factor level column in the result. Defaults to "Level".

na_string
String for NA factor levels in the result. Defaults to "(missing)".

strata_sep
Delimiter to separate stratification levels by in the result. Defaults to "/".

summary_strata_sep
Delimiter to separate summary column names with the strata groups. Defaults to "_".

fill_blanks
String to fill in blank spaces in the result. Defaults to "".

caption
Caption for resulting table passed to knitr::kable. Defaults to NULL.

...
Additional arguments to pass to "descriptives".

Value

A table of summary statistics according to the specified format. A tibble is returned if format = "none".

Details

The following statistics are available by default for each data type:

Numeric: "min", "max", "median", "iqr", "mean", "sd", "q1", "q3"

Categorical: "count", "percent"

All variables: "length", "missing", "available", "class", "unique"

These strings are typed explicitly in the ._summary arguments and serve as placeholders for where the actual value will appear. Custom functions can be entered in a named list, where the names are what provide access to the values in string templates. See "descriptives" and "absorb".

The names of the ._summary arguments are what become the column headers in the result. If unnamed, an arbitrary name (i.e. "VX") will appear in the column header.

Examples

require(tidyverse)

#1) Default summary
heart_disease %>%
    univariate_table()

#> <table class="table table-striped table-responsive" style="width: auto !important; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Variable </th>
#>    <th style="text-align:left;"> Level </th>
#>    <th style="text-align:left;"> Summary </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> Age </td>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:left;"> 56 (48, 61) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;vertical-align: top !important;" rowspan="2"> Sex </td>
#>    <td style="text-align:left;"> Female </td>
#>    <td style="text-align:left;"> 97 (32.01%) </td>
#>   </tr>
#>   <tr>
#>    
#>    <td style="text-align:left;"> Male </td>
#>    <td style="text-align:left;"> 206 (67.99%) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;vertical-align: top !important;" rowspan="4"> ChestPain </td>
#>    <td style="text-align:left;"> Typical angina </td>
#>    <td style="text-align:left;"> 23 (7.59%) </td>
#>   </tr>
#>   <tr>
#>    
#>    <td style="text-align:left;"> Atypical angina </td>
#>    <td style="text-align:left;"> 50 (16.5%) </td>
#>   </tr>
#>   <tr>
#>    
#>    <td style="text-align:left;"> Non-anginal pain </td>
#>    <td style="text-align:left;"> 86 (28.38%) </td>
#>   </tr>
#>   <tr>
#>    
#>    <td style="text-align:left;"> Asymptomatic </td>
#>    <td style="text-align:left;"> 144 (47.52%) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> BP </td>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:left;"> 130 (120, 140) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Cholesterol </td>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:left;"> 241 (211, 275) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> BloodSugar </td>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:left;"> 2 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> MaximumHR </td>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:left;"> 153 (133.5, 166) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;vertical-align: top !important;" rowspan="2"> ExerciseInducedAngina </td>
#>    <td style="text-align:left;"> No </td>
#>    <td style="text-align:left;"> 204 (67.33%) </td>
#>   </tr>
#>   <tr>
#>    
#>    <td style="text-align:left;"> Yes </td>
#>    <td style="text-align:left;"> 99 (32.67%) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;vertical-align: top !important;" rowspan="2"> HeartDisease </td>
#>    <td style="text-align:left;"> No </td>
#>    <td style="text-align:left;"> 164 (54.13%) </td>
#>   </tr>
#>   <tr>
#>    
#>    <td style="text-align:left;"> Yes </td>
#>    <td style="text-align:left;"> 139 (45.87%) </td>
#>   </tr>
#> </tbody>
#> </table>

#2) Stratified summary
heart_disease %>%
    univariate_table(
        strata = ~Sex,
        add_n = TRUE
    )

#> <table class="table table-striped table-responsive" style="width: auto !important; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Variable </th>
#>    <th style="text-align:left;"> Level </th>
#>    <th style="text-align:left;"> Female (N=97) </th>
#>    <th style="text-align:left;"> Male (N=206) </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> Age </td>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:left;"> 57 (50, 63) </td>
#>    <td style="text-align:left;"> 54.5 (47, 59.75) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;vertical-align: top !important;" rowspan="4"> ChestPain </td>
#>    <td style="text-align:left;"> Typical angina </td>
#>    <td style="text-align:left;"> 4 (4.12%) </td>
#>    <td style="text-align:left;"> 19 (9.22%) </td>
#>   </tr>
#>   <tr>
#>    
#>    <td style="text-align:left;"> Atypical angina </td>
#>    <td style="text-align:left;"> 18 (18.56%) </td>
#>    <td style="text-align:left;"> 32 (15.53%) </td>
#>   </tr>
#>   <tr>
#>    
#>    <td style="text-align:left;"> Non-anginal pain </td>
#>    <td style="text-align:left;"> 35 (36.08%) </td>
#>    <td style="text-align:left;"> 51 (24.76%) </td>
#>   </tr>
#>   <tr>
#>    
#>    <td style="text-align:left;"> Asymptomatic </td>
#>    <td style="text-align:left;"> 40 (41.24%) </td>
#>    <td style="text-align:left;"> 104 (50.49%) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> BP </td>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:left;"> 132 (120, 140) </td>
#>    <td style="text-align:left;"> 130 (120, 140) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Cholesterol </td>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:left;"> 254 (215, 302) </td>
#>    <td style="text-align:left;"> 235 (208.75, 268.5) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> BloodSugar </td>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:left;"> 2 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> MaximumHR </td>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:left;"> 157 (142, 165) </td>
#>    <td style="text-align:left;"> 150.5 (132, 167.5) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;vertical-align: top !important;" rowspan="2"> ExerciseInducedAngina </td>
#>    <td style="text-align:left;"> No </td>
#>    <td style="text-align:left;"> 75 (77.32%) </td>
#>    <td style="text-align:left;"> 129 (62.62%) </td>
#>   </tr>
#>   <tr>
#>    
#>    <td style="text-align:left;"> Yes </td>
#>    <td style="text-align:left;"> 22 (22.68%) </td>
#>    <td style="text-align:left;"> 77 (37.38%) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;vertical-align: top !important;" rowspan="2"> HeartDisease </td>
#>    <td style="text-align:left;"> No </td>
#>    <td style="text-align:left;"> 72 (74.23%) </td>
#>    <td style="text-align:left;"> 92 (44.66%) </td>
#>   </tr>
#>   <tr>
#>    
#>    <td style="text-align:left;"> Yes </td>
#>    <td style="text-align:left;"> 25 (25.77%) </td>
#>    <td style="text-align:left;"> 114 (55.34%) </td>
#>   </tr>
#> </tbody>
#> </table>