Compute descriptive statistics on columns of a data frame

The user can specify an unlimited number of functions to evaluate and the types of data that each set of functions will be applied to (including the default; see "Details").

descriptives(
    data,
    f_all = NULL,
    f_numeric = NULL,
    numeric_types = "numeric",
    f_categorical = NULL,
    categorical_types = "factor",
    f_other = NULL,
    useNA = c("ifany", "no", "always"),
    round = 2,
    na_string = "(missing)"
)

Arguments

data
A base::data.frame.

f_all
A base::list of functions to evaluate on all columns.

f_numeric
A base::list of functions to evaluate on numeric_types columns.

numeric_types
Character vector of data types that should be evaluated by f_numeric.

f_categorical
A base::list of functions to evaluate on categorical_types columns.

categorical_types
Character vector of data types that should be evaluated by f_categorical.

f_other
A base::list of functions to evaluate on remaining columns.

useNA
See base::table for details. Defaults to "ifany".

round
Digit to round numeric data. Defaults to 2.

na_string
String to fill in NA names.

Details

The min, max, median, iqr, mean, sd are automatically computed for numeric data and table, prop.table*100 for categorical data. The sample size, number of missing values, number of nonmissing values, the number of unique values, the class are automatically computed on all columns.

Value

A tibble::tibble with the following columns:

column: Position/index of the input dataset

column_name: Label of the column

type: Set of column types function was applied to

key: Name of function that was evaluated

value_order: Position of the value within the function result

value_name: Label extracted from the result with base::names

value_num: Numeric results

value_chr: Non-numeric results

value_combo: Output combined with rounding applied to numeric values

Examples

#Default
heart_disease %>%
    descriptives()

#> # A tibble: 111 x 9
#>    column column_name type  key   value_order value_name value_num value_chr
#>     <int> <chr>       <chr> <chr>       <int> <chr>          <dbl> <chr>    
#>  1      1 Age         all   avai…           1 NA             303   NA       
#>  2      1 Age         all   class           1 NA              NA   numeric  
#>  3      1 Age         all   leng…           1 NA             303   NA       
#>  4      1 Age         all   miss…           1 NA               0   NA       
#>  5      1 Age         all   uniq…           1 NA              41   NA       
#>  6      1 Age         nume… iqr             1 NA              13   NA       
#>  7      1 Age         nume… max             1 NA              77   NA       
#>  8      1 Age         nume… mean            1 NA              54.4 NA       
#>  9      1 Age         nume… medi…           1 NA              56   NA       
#> 10      1 Age         nume… min             1 NA              29   NA       
#> # … with 101 more rows, and 1 more variable: value_combo <chr>

#Allow logicals as categorical
heart_disease %>%
    descriptives(
        categorical_types = c("logical", "factor")
    )

#> # A tibble: 117 x 9
#>    column column_name type  key   value_order value_name value_num value_chr
#>     <int> <chr>       <chr> <chr>       <int> <chr>          <dbl> <chr>    
#>  1      1 Age         all   avai…           1 NA             303   NA       
#>  2      1 Age         all   class           1 NA              NA   numeric  
#>  3      1 Age         all   leng…           1 NA             303   NA       
#>  4      1 Age         all   miss…           1 NA               0   NA       
#>  5      1 Age         all   uniq…           1 NA              41   NA       
#>  6      1 Age         nume… iqr             1 NA              13   NA       
#>  7      1 Age         nume… max             1 NA              77   NA       
#>  8      1 Age         nume… mean            1 NA              54.4 NA       
#>  9      1 Age         nume… medi…           1 NA              56   NA       
#> 10      1 Age         nume… min             1 NA              29   NA       
#> # … with 107 more rows, and 1 more variable: value_combo <chr>

#Only apply "other" functions to numeric types
heart_disease %>%
    descriptives(
        numeric_types = NULL
    )

#> # A tibble: 75 x 9
#>    column column_name type  key   value_order value_name value_num value_chr
#>     <int> <chr>       <chr> <chr>       <int> <chr>          <dbl> <chr>    
#>  1      1 Age         all   avai…           1 NA               303 NA       
#>  2      1 Age         all   class           1 NA                NA numeric  
#>  3      1 Age         all   leng…           1 NA               303 NA       
#>  4      1 Age         all   miss…           1 NA                 0 NA       
#>  5      1 Age         all   uniq…           1 NA                41 NA       
#>  6      2 Sex         all   avai…           1 NA               303 NA       
#>  7      2 Sex         all   class           1 NA                NA factor   
#>  8      2 Sex         all   leng…           1 NA               303 NA       
#>  9      2 Sex         all   miss…           1 NA                 0 NA       
#> 10      2 Sex         all   uniq…           1 NA                 2 NA       
#> # … with 65 more rows, and 1 more variable: value_combo <chr>

#Compute a custom function
heart_disease %>%
    descriptives(
        f_numeric =
            list(
                cv = function(x) sd(x, na.rm = TRUE)/mean(x, na.rm = TRUE)
            )
    )

#> # A tibble: 115 x 9
#>    column column_name type  key   value_order value_name value_num value_chr
#>     <int> <chr>       <chr> <chr>       <int> <chr>          <dbl> <chr>    
#>  1      1 Age         all   avai…           1 NA           303     NA       
#>  2      1 Age         all   class           1 NA            NA     numeric  
#>  3      1 Age         all   leng…           1 NA           303     NA       
#>  4      1 Age         all   miss…           1 NA             0     NA       
#>  5      1 Age         all   uniq…           1 NA            41     NA       
#>  6      1 Age         nume… cv              1 NA             0.166 NA       
#>  7      1 Age         nume… iqr             1 NA            13     NA       
#>  8      1 Age         nume… max             1 NA            77     NA       
#>  9      1 Age         nume… mean            1 NA            54.4   NA       
#> 10      1 Age         nume… medi…           1 NA            56     NA       
#> # … with 105 more rows, and 1 more variable: value_combo <chr>