Stretch one or variables over many columns by one or more keys

Provides similar functionality to tidyr::spread or reshape2::dcast but allows for an unlimited number of variables to be spanned across the columns.

stretch(
    data,
    keys,
    keep = NULL,
    send = NULL,
    join = dplyr::full_join,
    .sep = "_",
    extract_keys_as_header = FALSE,
    keep_keys_in_header = TRUE,
    ...
)

Arguments

data
A data.frame object.

keys
A vector of column names or positions, or a tidyselect::vars_select helper.

keep
Variables to remain as their own columns. If NULL (default), all variables except those in send are chosen, excluding the keys. Has tidyselect::vars_select capabilities.

send
Variables to send into new columns for each key. If NULL (default), all variables except those in keep are chosen, excluding the keys. Has tidyselect::vars_select capabilities.

join
Function that joins on keep across the keys. Defaults to dplyr::full_join. See ?dplyr::join for choices.

.sep
Character to separate and identify keys values over the columns. Defaults to "_".

extract_keys_as_header
Should the keys labels be returned as a separate character vector? Defaults to FALSE. Has no effect when there is only 1 send column.

keep_keys_in_header
If extract_keys_as_header = TRUE, should the keys be left in the result columns? Defaults to TRUE. Useful to set to FALSE if a call to knitr::kable follows.

...
Additional arguments to be passed to join.

Value

A wide data.frame with variables spread over the columns. If extract_keys_as_header = TRUE, the result is a two-element list with the transformed data in 1 element and the top-level header in the other. If only 1 send variable exists, the original names are removed.

Examples

require(tidyverse)

#Make data frame with multiple summary columns 
temp_summary <-

    heart_disease %>%
    group_by(
        Sex,
        HeartDisease,
        BloodSugar
    ) %>%
    summarise(
        Mean = mean(Age, na.rm = TRUE),
        SD = sd(Age, na.rm = TRUE),
        Median = median(Age, na.rm = TRUE)
    ) %>%
    ungroup()

#1) Span summaries for each combination of Sex and BloodSugar
temp_summary %>%
    stretch(
        keys = c("Sex", "BloodSugar"),
        keep = "HeartDisease"
    )

#> # A tibble: 2 x 13
#>   HeartDisease Mean_Female_FAL… SD_Female_FALSE Median_Female_F… Mean_Male_FALSE
#>   <fct>                   <dbl>           <dbl>            <dbl>           <dbl>
#> 1 No                       54.0           10.4              53.5            50.2
#> 2 Yes                      59.7            3.61             61              55.8
#> # … with 8 more variables: SD_Male_FALSE <dbl>, Median_Male_FALSE <dbl>,
#> #   Mean_Female_TRUE <dbl>, SD_Female_TRUE <dbl>, Median_Female_TRUE <dbl>,
#> #   Mean_Male_TRUE <dbl>, SD_Male_TRUE <dbl>, Median_Male_TRUE <dbl>

#2) If "HeartDisease" wasn't fully crossed, use different joining to get only matching groups
temp_summary %>%
    stretch(
        keys = c("Sex", "BloodSugar"),
        keep = "HeartDisease",
        join = inner_join
    )

#> # A tibble: 2 x 13
#>   HeartDisease Mean_Female_FAL… SD_Female_FALSE Median_Female_F… Mean_Male_FALSE
#>   <fct>                   <dbl>           <dbl>            <dbl>           <dbl>
#> 1 No                       54.0           10.4              53.5            50.2
#> 2 Yes                      59.7            3.61             61              55.8
#> # … with 8 more variables: SD_Male_FALSE <dbl>, Median_Male_FALSE <dbl>,
#> #   Mean_Female_TRUE <dbl>, SD_Female_TRUE <dbl>, Median_Female_TRUE <dbl>,
#> #   Mean_Male_TRUE <dbl>, SD_Male_TRUE <dbl>, Median_Male_TRUE <dbl>

#3) Only send two of the summaries
temp_summary %>%
    stretch(
        keys = c("Sex", "BloodSugar"),
        keep = "HeartDisease",
        send = c("Mean", "Median")
    )

#> # A tibble: 2 x 9
#>   HeartDisease Mean_Female_FAL… Median_Female_F… Mean_Male_FALSE
#>   <fct>                   <dbl>            <dbl>           <dbl>
#> 1 No                       54.0             53.5            50.2
#> 2 Yes                      59.7             61              55.8
#> # … with 5 more variables: Median_Male_FALSE <dbl>, Mean_Female_TRUE <dbl>,
#> #   Median_Female_TRUE <dbl>, Mean_Male_TRUE <dbl>, Median_Male_TRUE <dbl>

#4) Clean HTML table with keys spanned over columns
result <-
    temp_summary %>%
    stretch(
        keys = c("Sex", "BloodSugar"),
        keep = "HeartDisease",
        extract_keys_as_header = TRUE,
        keep_keys_in_header = FALSE
    )
result$.result %>%
    knitr::kable(format = "html") %>%
    kableExtra::kable_styling() %>%
    kableExtra::add_header_above(
        kableExtra::auto_index(result$.header)
    )

#> <table class="table" style="margin-left: auto; margin-right: auto;">
#>  <thead>
#> <tr>
#> <th style="border-bottom:hidden" colspan="1"></th>
#> <th style="border-bottom:hidden; padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="3"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Female_FALSE</div></th>
#> <th style="border-bottom:hidden; padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="3"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Male_FALSE</div></th>
#> <th style="border-bottom:hidden; padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="3"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Female_TRUE</div></th>
#> <th style="border-bottom:hidden; padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="3"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Male_TRUE</div></th>
#> </tr>
#>   <tr>
#>    <th style="text-align:left;"> HeartDisease </th>
#>    <th style="text-align:right;"> Mean </th>
#>    <th style="text-align:right;"> SD </th>
#>    <th style="text-align:right;"> Median </th>
#>    <th style="text-align:right;"> Mean </th>
#>    <th style="text-align:right;"> SD </th>
#>    <th style="text-align:right;"> Median </th>
#>    <th style="text-align:right;"> Mean </th>
#>    <th style="text-align:right;"> SD </th>
#>    <th style="text-align:right;"> Median </th>
#>    <th style="text-align:right;"> Mean </th>
#>    <th style="text-align:right;"> SD </th>
#>    <th style="text-align:right;"> Median </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> No </td>
#>    <td style="text-align:right;"> 54.03030 </td>
#>    <td style="text-align:right;"> 10.408532 </td>
#>    <td style="text-align:right;"> 53.5 </td>
#>    <td style="text-align:right;"> 50.21333 </td>
#>    <td style="text-align:right;"> 8.819961 </td>
#>    <td style="text-align:right;"> 51.0 </td>
#>    <td style="text-align:right;"> 60.33333 </td>
#>    <td style="text-align:right;"> 6.653320 </td>
#>    <td style="text-align:right;"> 59 </td>
#>    <td style="text-align:right;"> 54.70588 </td>
#>    <td style="text-align:right;"> 6.761700 </td>
#>    <td style="text-align:right;"> 53 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Yes </td>
#>    <td style="text-align:right;"> 59.68421 </td>
#>    <td style="text-align:right;"> 3.606362 </td>
#>    <td style="text-align:right;"> 61.0 </td>
#>    <td style="text-align:right;"> 55.75510 </td>
#>    <td style="text-align:right;"> 8.560351 </td>
#>    <td style="text-align:right;"> 57.5 </td>
#>    <td style="text-align:right;"> 57.16667 </td>
#>    <td style="text-align:right;"> 7.808115 </td>
#>    <td style="text-align:right;"> 58 </td>
#>    <td style="text-align:right;"> 58.12500 </td>
#>    <td style="text-align:right;"> 7.116881 </td>
#>    <td style="text-align:right;"> 58 </td>
#>   </tr>
#> </tbody>
#> </table>